requirements.txt
pytest
requests
influxdb-client
python-dotenv